HW 7 Fixes
