+++ create-player.js
@@ -1,0 +1,22 @@
+import { createElement } from "./create-element.js";
+
+export function createPlayer(playerInfo) {
+    const $player = createElement('div', 'player'+playerInfo.player);
+    const $progressbar = createElement('div', 'progressbar');
+    const $character = createElement('div', 'character');
+    const $life = createElement('div', 'life');
+    const $name = createElement('div', 'name');
+    const $img = createElement('img');
+
+    $life.style.width = playerInfo.hp + '%';
+    $name.innerText = playerInfo.name;
+    $img.src = playerInfo.img;
+
+    $player.appendChild($progressbar);
+    $player.appendChild($character);
+    $progressbar.appendChild($life);
+    $progressbar.appendChild($name);
+    $character.appendChild($img);
+
+    return $player;
+}
