+++ elements.js
@@ -1,4 +1,6 @@
-export const formFight = document.querySelector('.control');
-export const arenas = document.querySelector('.arenas');
-export const fightButton = formFight.querySelector('.button');
-export const chat = document.querySelector('.chat');
+const formFight = document.querySelector('.control');
+const arenas = document.querySelector('.arenas');
+const fightButton = formFight.querySelector('.button');
+const chat = document.querySelector('.chat');
+
+export { formFight, arenas, fightButton, chat };
