+++ values.js
@@ -1,7 +1,9 @@
-export const HIT = {
+const HIT = {
     head: 30,
     body: 25,
     foot: 20,
 };
 
-export const ATTACK = ['head', 'body', 'foot'];
+const ATTACK = ['head', 'body', 'foot'];
+
+export { HIT, ATTACK };
